Make full stacktraces hidden by default (initial attempt)
A very rough around the edges attempt just to get a feel for what I'm
doing.  Key highlights for discussion:

* Add jQuery library (writing this in raw js is not fun, with jQuery
it's just a few lines to implement)
* jQuery code injected in AbstractHtmlCreator

I imagine you're not a big fan of injecting jQuery, but I can't bring
myself to write raw js when jQuery takes a fraction of the time and a
fraction of the code.

Injecting the script at AbstractHtmlCreator seems a bit hackish, but
maybe not?  Not sure if you'd want to create a new entry point to do it?
If it's not jquery, some kind of javascript needs to be injected
somewhere/somehow to make this happen.

The links are just the first line of the stacktrace, which isn't pretty,
and not ideal.  If this implmenetation seems like it will be accepted,
I'll work on making the links a little more user friendly (i.e. display
the entire exception.getMessage() string), but for now it gives you an
idea of what I'm thinking of with this.

Unit tests will fail due to injected jQuery and js into templates.